+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Added getter of tracked jobs list.
 - Added remove worker feature.
 
 ## [3.7.0] - 2019-04-19